收紧 Go 桌面客户端文件确认弹窗节流

+++ desktop-client-go/internal/app/clipboard_toast_windows.go
@@ -36,7 +36,7 @@ func (a *App) showClipboardFilesToast(paths []string, windowSeconds int) bool {
 		Actions: []toast.Action{
 			{Type: toast.Protocol, Content: "立即发送", Arguments: confirmURL},
 		},
-		Duration: toast.Long,
+		Duration: toast.Short,
 	}
 	if strings.TrimSpace(panelURL) != "" {
 		notification.Actions = append(notification.Actions, toast.Action{
